% Auto-generated: do not edit by hand
\name{linearHistComponent}

\alias{linearHistComponent}

\title{LinearHistComponent component}

\description{

}

\usage{
linearHistComponent(children=NULL, id=NULL, data=NULL, histData=NULL,
histValue=NULL, value=NULL, x_axis=NULL, y_axis=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{data}{List of logical | numeric | character | named list | unnamed lists. Update props to trigger callbacks.}

\item{histData}{List of logical | numeric | character | named list | unnamed lists. }

\item{histValue}{Character. }

\item{value}{Character. }

\item{x_axis}{Character. }

\item{y_axis}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

