% Auto-generated: do not edit by hand
\name{myDashComponents}

\alias{myDashComponents}

\title{MyDashComponents component}

\description{

}

\usage{
myDashComponents(children=NULL, id=NULL, data=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{data}{List of logical | numeric | character | named list | unnamed lists. Update props to trigger callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

